begin refactoring of Akkadian
